Move WPF Polygon3D tests
