+++ Source/HelixToolkit.Wpf.Tests/HelixToolkit.Wpf.Tests.csproj
@@ -100,7 +100,6 @@
     <Compile Include="Importers\StlReaderTests.cs" />
     <Compile Include="Importers\ObjReaderTests.cs" />
     <Compile Include="Importers\LwoReaderTests.cs" />
-    <Compile Include="Polygon3DTests.cs" />
     <Compile Include="StereoHelperTests.cs" />
     <Compile Include="Properties\AssemblyDescription.cs" />
     <Compile Include="Visual3Ds\MeshVisuals\CubeVisual3DTests.cs" />
